Ensure that config.h (which contains the results of all of the configure tests) is included in each source file of the library (#1)
Because we have specified AC_CONFIG_HEADER() in configure.ac, we are telling autoconf and autoheader that we support the use config.h instead of requiring the feature test macros to be passed as -D... arguments on the compiler command-line - in exchange, autoconf tells us that it is indeed using config.h by passing -DHAVE_CONFIG_H on the compiler's command-line.

Therefore, if we care about anything tested by the configure script, we should always include the stanza in this commit at the top of each source file.