Remove attributes.h (#1)
This header was used across both public (installed) and private sources, which means that an application attempting to include <libawsclient.h> would fail to compile. We can't go around installing a file named something generic such as "attributes.h" to $(includedir), and so the simplest path is to merge its contents into libawsclient.h itself.